Fix library updating
Cards are enumerated with a number value to match against the todo
library when populating the modals.

Before if you clicked on a project on sidebar to load that list,
they would be displayed, but if you then clicked on either the edit
button or details button, it would populate the modal based on the
numerations according to the whole todo library.
as opposed to according to the numeration of the list of that specific
group list.

That is now fixed by sending what specific library is called when
sidebar is clicked